+++ react-components/src/pages/Main/components/Card/index.scss
