+++ react-components/src/pages/Main/components/Cards/index.scss
