+++ react-components/src/pages/Main/components/ModalPopup/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import { useCardId } from '../../hooks/useCardId';
+import { useCardId } from '../../../../hooks/useCardId';
 import ModalWindow from '../ModalWindow';
-import { useGetCardQuery } from '../../store/api/cards.api';
+import { useGetCardQuery } from '../../../../store/api/cards.api';
 
 const ModalPopup = () => {
   const cardId = useCardId();
