+++ react-components/src/pages/Main/components/ModalWindow/index.test.tsx
