+++ react-components/src/pages/Main/components/ModalWindow/index.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
-import { Card as ICard } from '../../types/card';
+import { Card as ICard } from '../../../../types/card';
 import { useSearchParams } from 'react-router-dom';
 import './index.scss';
-import Loader from '../Loader';
-import { useActions } from '../../hooks/useAction';
+import Loader from '../../../../components/Loader';
+import { useActions } from '../../../../hooks/useAction';
 
 interface ModalWindowProps {
   modalData: ICard[] | null;
