+++ react-components/src/pages/Main/components/Search/index.scss
