+++ react-components/src/pages/Main/components/Search/index.test.tsx
