+++ react-components/src/pages/Main/index.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
-import Search from 'components/Search';
-import Cards from '../../components/Cards';
+import Search from 'pages/Main/components/Search';
+import Cards from './components/Cards';
 
 import Loader from '../../components/Loader';
 import { useGetCardsQuery } from '../../store/api/cards.api';
 
 import { useSearch } from '../../hooks/useSearch';
 import { useCardId } from '../../hooks/useCardId';
-import ModalPopup from '../../components/ModalPopup';
+import ModalPopup from './components/ModalPopup';
 
 const Main = () => {
   const search = useSearch();
